laravel conversion, let see how it goes
